+++ persistence-mongo/src/main/java/de/evoila/cf/broker/persistence/repository/ServiceDefinitionRepositoryImpl.java
@@ -4,8 +4,8 @@
 package de.evoila.cf.broker.persistence.repository;
 
 import de.evoila.cf.broker.exception.ServiceDefinitionDoesNotExistException;
-import de.evoila.cf.broker.model.Plan;
-import de.evoila.cf.broker.model.ServiceDefinition;
+import de.evoila.cf.broker.model.catalog.plan.Plan;
+import de.evoila.cf.broker.model.catalog.ServiceDefinition;
 import de.evoila.cf.broker.repository.ServiceDefinitionRepository;
 import de.evoila.cf.broker.service.CatalogService;
 import org.springframework.stereotype.Repository;
